add keep images, add pause/resume, add manual refresh

+++ Downlink-win/Helpers/Scheduler.cs
@@ -25,11 +25,17 @@ namespace Downlink_win.Helpers
 
         public async Task RunScheduler()
         {
+            // reset interval
             schedulerInterval = MainWindow._localSettings.WallpaperSource.interval;
+
+            // set timer for next scheduler run
+            MainWindow._bindings.LastUpdated = $"Last updated {DateTime.Now}";
+            StartTimers(false);
 
             // download image, return its path, set as wallpaper and delete
             var imagePath = await _downloadHelper.DownloadAndSaveImage(MainWindow._localSettings.WallpaperSource.url.full);
 
+            // process downloaded image
             if (!imagePath.Any())
             {
                 MainWindow._bindings.ProgramStatus = "Download failed. Retrying shortly...";
@@ -38,19 +44,31 @@ namespace Downlink_win.Helpers
             else
             {
                 _wallpaperHelper.SetWallpaper(imagePath);
-                _wallpaperHelper.DeleteImage(imagePath);
 
-                MainWindow._bindings.LastUpdated = $"Last updated {DateTime.Now}";
-
-                StartTimers(false);
+                if (MainWindow._localSettings.KeepImages)
+                    _wallpaperHelper.MoveImage(imagePath);
+                else
+                    _wallpaperHelper.DeleteImage(imagePath);
             }
-            
+        }
+
+        public void PauseScheduler()
+        {
+            MainWindow._bindings.ProgramStatus = "Paused.";
+            schedulerTimer.Stop();
+        }
+
+        public void ResumeScheduler()
+        {
+            MainWindow._bindings.ProgramStatus = "";
+            schedulerTimer.Start();
         }
 
         private void StartTimers(bool fastRetry)
         {
             schedulerTimer.Stop();
 
+            // retry in 10s (in case of download failure) or the source's defined interval
             schedulerTimer.Interval = (fastRetry ? 10 : schedulerInterval) * 1000;
             schedulerTimer.Start();
         }
